+++ 10LaserDeLaJustice/src/interfaces/SceneListener.java
@@ -8,5 +8,5 @@ import java.util.EventListener;
 public interface SceneListener extends EventListener{
 	
 	public void couleurLaserListener();
-	public void changementTempsListener(int temps);
+	public void changementTempsListener(double temps);
 }
